TargetScanner and Gun rotation complete
PlayerGun became PlayerGunTargetScanner and created new Plaayergun
Targetscanner gets the target based on if it is in range, is there clear line of sight and if it's withing the contraint angle(like field of view of pilot if it were 3D but only top to bottom), then holds the target on a public property of type Transform. PlayerGun takes this transform and rotates the gun accordingly. Scanner only passes the target transform and nothing else. so that math issues in one doesn't affect the other.