Auto targeting player gun (#7)
* Scanning state implemented

* Scanning and tracking system done

Now we have to just get the target from it and use that to move the turret , keeping in mind the angle constraints

* Fixed the direction of raycast and raycast visualizing gizmo

* Now this class checks to see if target is in constraint

So gun moving code won't have to be mindful of that, can just go back to start rotation when target is lost

* TargetScanner and Gun rotation complete

PlayerGun became PlayerGunTargetScanner and created new Plaayergun
Targetscanner gets the target based on if it is in range, is there clear line of sight and if it's withing the contraint angle(like field of view of pilot if it were 3D but only top to bottom), then holds the target on a public property of type Transform. PlayerGun takes this transform and rotates the gun accordingly. Scanner only passes the target transform and nothing else. so that math issues in one doesn't affect the other.

* Gun complete